JEL-299: improve integration test tooling
